requirements.txt
cassandra-driver==3.30.0
click==8.3.3
Deprecated==1.3.1
Faker==40.15.0
geomet==1.1.0
python-dotenv==1.2.2
wrapt==2.1.2
